Move Github Event emitter Init
