+++ test/utils/github.ts
@@ -6,7 +6,7 @@ import merge from 'lodash.merge';
 
 import { Fastify } from '@types';
 
-import { githubEvents } from '@/api/github';
+import { githubEvents } from '@/init/github';
 
 type DeepPartial<T> = {
   [P in keyof T]?: T[P] extends (infer U)[]
